refactor: extract withSchema to db.ts; add UTC comment; fix missing release assert
- Move withSchema<T> from mcp.ts into db.ts so both query.ts and mcp.ts
  share the same connection-pinning helper (addresses review comment)
- query.ts now imports withSchema from db.ts instead of inlining the pattern
- mcp.ts removes its local withSchema definition and imports from db.ts
- Add clarifying comment on Date serialisation: .toISOString() is always UTC
  so displayed time is only correct when MySQL server is also UTC
- Add missing expect(mockConn.release).toHaveBeenCalledTimes(1) to the DML
  test case for consistency with all other test cases
- Update vi.mock factory to supply withSchema so connection-pinning
  assertions still pass after the extraction
